refactor(section): 주어진 위치에 구간을 추가하는 private 메서드를 추출했다.
